requirements.txt
asgiref==3.8.1
Django==5.1.1
django-js-asset==2.2.0
django-mptt==0.16.0
sqlparse==0.5.1
